refactor: useCouponStorage 훅 추가 및 App 컴포넌트에서 쿠폰 상태 관리 개선
